+++ packages/fleur-benchmarks/jest.config.js
@@ -1,13 +1,11 @@
-const { join } = require('path')
-
 module.exports = {
   ...require('../../jest.config'),
-  moduleFileExtensions: ['ts', 'js'],
+  moduleFileExtensions: ['ts', 'tsx', 'js'],
   setupFilesAfterEnv: ['<rootDir>/src/setup.ts'],
   globals: {
     'ts-jest': {
-      tsConfigFile: 'tsconfig.json',
-      typeCheck: false,
+      tsConfig: 'tsconfig.json',
+      isolatedModules: true,
     },
   },
 }