+++ packages/fleur/src/index.ts
@@ -5,8 +5,12 @@ export { StoreContext } from './StoreContext'
 export { OperationContext } from './OperationContext'
 export { action, actions, ActionsOf } from './Action'
 export { AppContext, StoreGetter } from './AppContext'
-export { ComponentContext } from './ComponentContext'
-export { operations, operation, Operation, OperationArgs } from './Operations'
+export {
+  operations,
+  operation,
+  OperationDef as Operation,
+  OperationArgs,
+} from './Operations'
 export { selector } from './selector'
 export { selectorWithStore } from './selectorWithStore'
 export { withReduxDevTools } from './withReduxDevtools'