+++ packages/testing/src/mockStore.spec.ts
@@ -1,6 +1,5 @@
 import { mockStore } from './mockStore'
 import { Store } from '@fleur/fleur'
-import { mockOperationContext } from './mockOperationContext'
 
 describe('mockStoreContext', () => {
   it('Should create Store entry correctly', () => {